Refactor: split Activity and SetList into separate components
