+++ src/feature/SetList/SetList.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import React from 'react';
+
+export const SetList = () => {
+  return (
+    <section className="mb-6">
+      <h2 className="text-3xl font-extrabold">SetList</h2>
+    </section>
+  );
+};
